requirements.txt
anthropic
pandas
numpy
ta
scipy
plotly
yfinance
streamlit
python-dotenv
openai
